test: Fix flaky test due to persistent environment variable
If the tests against the CLI are run together and the one setting the
environment variable runs first, the one supposed to run without the
environment variable will fail, since the environment variable will
remain set.

Using the `StartInfo` of the process, the environment variable is
limited to the dedicated process.

#SDPI-11